Clean up unused files, clang-format
